add some style

+++ public/javascripts/main.js
@@ -19,23 +19,26 @@ app.controller("BlogCtrl",['$http', function($http) {
 
 		$http.post('/api/posts', this.newPost).
 			success(function(){
-				that.posts.unshift(that.newPost)
-				that.newPost = {}
+				that.newPost.updatedAt = new Date();
+				that.posts.unshift(that.newPost);
+				that.newPost = {};
 			});
 	};
 
 	this.deletePostByID = function(id){
 
-		$http.delete('/api/posts/'+id).
-			success(function(){
+		if (confirm("确定删除？")){
+			$http.delete('/api/posts/'+id).
+				success(function(){
 
-				that.posts.forEach(function(val, ind){
-					if(val._id === id){
-						that.posts.splice(ind, 1);
-					}
+					that.posts.forEach(function(val, ind){
+						if(val._id === id){
+							that.posts.splice(ind, 1);
+						}
+					});
+
 				});
-
-			});
+		}
 
 	};
